Add first jumbotron image and content;
add a Bootstrap "jumbotron" section. Use a <div> tag with two classes: jumbotron and jumbotron-bike.

In style.css, add ../images/jumbotron-bike.jpg as the background image for this section.
In this section in index.html, add the header text Get Going. Below that, add the following paragraph:
"Jumpstart your startup with some crowdsourced seed money, or support one of our 300,000,000+ creators and get rewards."
You will need the text-shadow property to achieve the outlines of the fonts.