Added slick carousel.
In main.js, when the document is ready, apply this configuration to the carousel:

One slide should scroll at a time.
The carousel should autoplay when the page is loaded.
The speed should be 4000 milliseconds.
Do not include arrow buttons.
Do not allow the user to drag the images.
Do not pause on "focus" or "hover".